+++ src/server/.gitignore
@@ -1,3 +1,4 @@
 .vscode
 node_modules
 temp.js
+*.log
