+++ src/server/api/package.json
@@ -17,6 +17,7 @@
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "mongoose": "^5.5.4",
-    "mosca": "^2.8.3"
+    "mosca": "^2.8.3",
+    "winston": "^3.2.1"
   }
 }
